% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createListOfMsExperimentFromListOfSpectra.R
\name{createListOfMsExperimentFromListOfSpectra}
\alias{createListOfMsExperimentFromListOfSpectra}
\title{Create list of \code{MsExperiment} objects from list of 
\code{Spectra} objects}
\usage{
createListOfMsExperimentFromListOfSpectra(sps_l)
}
\arguments{
\item{sps_l}{\code{list} of one or multiple \code{Spectra} objects}
}
\description{
The function \code{createListOfMsExperimentFromListOfSpectra} will 
provide the input (after subsetting the resulting list) for the 
\code{MsQuality} workflow for QC metric calculation.

The function \code{createListOfMsExperimentFromListOfSpectra} returns a 
list of \code{MsExperiment} objects. It will take as input a list of 
\code{Spectra} objects.
}
\details{
When \code{type == "MetIDQ"}, the function will return a list of length 2. 
The first entry contains a \code{MsExperiment} object with all features found 
in the experiment files. The second entry contains a \code{MsExperiment} 
object with the features found in the experiment files that have available 
information on their retention time.

When \code{type == "mzML"}, the function will return a lit of length 1. The
entry contains a \code{MsExperiment} object derived from all mzML files in 
\code{path}.
}
\examples{
## type == "MetIDQ"
path <- system.file("metidq", package = "MsQualityUtils")
sps_l <- createListOfSpectra(type = "MetIDQ", path = path, sheet = 1)
createListOfMsExperimentFromListOfSpectra(sps_l = sps_l)

## type == "mzML"
path <- system.file("sciex", package = "msdata")
sps_l <- createListOfSpectra(type = "mzML", path = path)
createListOfMsExperimentFromListOfSpectra(sps_l = sps_l)

}
\author{
Thomas Naake
}
