% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createListOfSpectra.R
\name{createListOfSpectra}
\alias{createListOfSpectra}
\title{Create list of Spectra object(s)}
\usage{
createListOfSpectra(type = c("MetIDQ", "mzML"), ...)
}
\arguments{
\item{type}{\code{character}, either \code{"MetIDQ"} or \code{"mzML"}}

\item{...}{arguments passed to 
\code{createListOfSummarizedExperimentFromMetIDQ}
or \code{createSpectraFromMzML}, for instance \code{path}}
}
\value{
list of \code{Spectra} object(s)
}
\description{
The function \code{createListOfSpectra} returns a list of \code{Spectra} 
objects.

When \code{type == "MetIDQ"}, the function will return a list of length 2. 
The first entry contains a \code{Spectra} object with all features found in 
the experiment files. The second entry contains a \code{Spectra} object with 
the features found in the experiment files that have available information on 
their retention time.

When \code{type == "mzML"}, the function will return a lit of length 1. The
entry contains a \code{Spectra} object derived from all mzML files in 
\code{path}.
}
\details{
If \code{type == "MetIDQ"}, the argument \code{rt} can be passed to the 
function. \code{rt} has to be a \code{SummarizedExperiment} object 
containing retention time values in the assay slot. By default, 
\code{createListOfSpectra} loads
an in-house library of retention time values from the metabolites of the
Biocrates MxP Quant 500 kit that are separated by liquid chromatography. 
The \code{Spectra} object only contains the the shared features between
\code{se} and \code{rt}.
}
\examples{
## type == "MetIDQ"
path <- system.file("metidq", package = "MsQualityUtils")
createListOfSpectra(type = "MetIDQ", path = path, sheet = 1)

## type == "mzML"
path <- system.file("sciex", package = "msdata")
createListOfSpectra(type = "mzML", path = path)
}
\author{
Thomas Naake
}
