% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraFromMetIDQ.R
\name{createListOfSummarizedExperimentFromMetIDQ}
\alias{createListOfSummarizedExperimentFromMetIDQ}
\title{Create list of SummarizedExperiment object from MetIDQ files}
\usage{
createListOfSummarizedExperimentFromMetIDQ(path = "./", sheet = 1, ...)
}
\arguments{
\item{path}{\code{character}}

\item{sheet}{\code{character} or \code{numeric}, the name or index to read 
data from}

\item{...}{additional parameters given to \code{read.xlsx}}
}
\description{
Load the xlsx files and store them as a SummarizedExperiment object in a 
list.
}
\details{
For each xlsx file found in path a new list entry will be created.
}
\examples{
path <- system.file("metidq", package = "MsQualityUtils")
createListOfSummarizedExperimentFromMetIDQ(path = path, sheet = 1)
}
\author{
Thomas Naake
}
