% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraFromMetIDQ.R
\name{createSpectraFromMetIDQ}
\alias{createSpectraFromMetIDQ}
\title{Create Spectra object from a list of SummarizedExperiment objects}
\usage{
createSpectraFromMetIDQ(se_l, rt)
}
\arguments{
\item{se_l}{list of \code{SummarizedExperiment} objects}

\item{rt}{\code{SummarizedExperiment} object containing the retention time 
in the assay slot}
}
\description{
The function \code{createSpectraFromMetIDQ} will a \code{Spectra} object
from list of \code{SummarizedExperiment} objects.
}
\details{
\code{createSpectraFromMetIDQ} will stop if there are duplicate sample 
names across the \code{SummarizedExperiment} objects.
}
\examples{
## create rt object
rt <- loadRt()

## create se object
path <- system.file("metidq", package = "MsQualityUtils")
se_l <- createListOfSummarizedExperimentFromMetIDQ(path = path, sheet = 1)

## run the createSpectraFromMetIDQ
createSpectraFromMetIDQ(se_l = se_l, rt = rt)
}
\author{
Thomas Naake
}
