% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraFromMzML.R
\name{createSpectraFromSrmOrMrmMzML}
\alias{createSpectraFromSrmOrMrmMzML}
\title{Create Spectra object from mzML files (SRM or MRM technology)}
\usage{
createSpectraFromSrmOrMrmMzML(path = ".")
}
\arguments{
\item{path}{\code{character}, path to mzML files}
}
\value{
\code{Spectra} object
}
\description{
The function \code{createSpectraFromSrmOrMrmMzML} creates a (single) 
\code{Spectra} object from all mzML files found in \code{path}.
}
\details{
Depending on the technology the quantitative information will 
be stored in different lists within the mzML files. In case of 
selected reaction monitoring (SRM)- or multiple reaction monitoring 
(MRM)-derived files, the intensities and m/z values are stored in 
\code{chromatogram} entries within \code{chromatogramList}.
The mzML files will be loaded via \code{MSnbase::readSRMData} and 
stored within a \code{MChromatogram} object. 

The entries in the returned \code{Spectra} object are separable by 
\code{Spectra$dataOrigin}.
}
\examples{
path <- system.file("srm", package = "MsQualityUtils")
createSpectraFromSrmOrMrmMzML(path = path)
}
\author{
Thomas Naake
}
