% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraFromMetIDQ.R
\name{createSpectraFromSummarizedExperiment}
\alias{createSpectraFromSummarizedExperiment}
\title{Create Spectra object from a SummarizedExperiment object}
\usage{
createSpectraFromSummarizedExperiment(se, rt)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object}

\item{rt}{\code{SummarizedExperiment} object containing the retention time 
in the assay slot}
}
\description{
Function to create a \code{Spectra} object from a \code{SummarizedExperiment}
object.
}
\details{
The sample name (\code{colnames(se)}) is taken to create the 
\code{dataOrigin} that is in down-stream functions used to distinguish the 
origins of the resulting \code{Spectra}'s entries.
}
\examples{
## create rt object
rt <- loadRt()

## create se object
path <- system.file("metidq", package = "MsQualityUtils")
se_l <- createListOfSummarizedExperimentFromMetIDQ(path = path, sheet = 1)
se <- se_l[[1]]

## run createSpectraFromSummarizedExperiment
createSpectraFromSummarizedExperiment(se = se, rt = rt)
}
\author{
Thomas Naake
}
