% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Retention_times_MetIDS-data.R
\docType{data}
\name{Retention_times_MetIDQ}
\alias{Retention_times_MetIDQ}
\title{Data for \code{MsQualityUtils}: Retention time for MetIDQ}
\format{
\code{SummarizedExperiment}
}
\source{
library(MatrixQCvis)
rt <- biocrates(file = "Retention_times_MetIDQ.xlsx", sheet = "2021_06_15_RT")

## convert to seconds
rt <- MatrixQCvis:::updateSE(se = rt, assay(rt) * 60)

saveRDS(rt, "Retention_times_MetIDQ.RDS")
}
\value{
\code{SummarizedExperiment}
}
\description{
\code{Retention_times_MetIDQ} contains the retention time
in seconds of the metabolites for the Biocrates MxP 500 Quant Kit measured 
by the Metabolomics Core Technology Platorm, COS, University of Heidelberg
(by Hagen Gegner).
It will be used to create a `Spectra` list containing only the metabolites
that were separated by liquid chromatography.
}
\author{
Thomas Naake
}
