% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectraFromMzML.R
\name{createSpectraFromMzML}
\alias{createSpectraFromMzML}
\title{Create Spectra object from mzML files}
\usage{
createSpectraFromMzML(path = ".")
}
\arguments{
\item{path}{\code{character}}
}
\value{
\code{Spectra} object
}
\description{
The function \code{createSpectraFromMzML} creates a (single) \code{Spectra} 
object from all mzML files found in \code{path}.
}
\details{
The package uses the \code{MsBackendMzR} for importing mzML files. The entries 
in the returned \code{Spectra} object are separable by 
\code{Spectra$dataOrigin}.
}
\examples{
path <- system.file("sciex", package = "msdata")
createSpectraFromMzML(path = path)
}
\author{
Thomas Naake
}
